- Modify the show application action and view
Previously, in the interview timeline for every interview details fetch a seperate
request is made. Currently, the logic is removed and all the interview details and its comments
are built in the same dom and sent as a single html file.

This should be changed later using ajax

Also, showing when the application page is loaded, the last interview round will be opened by default.